requirements.txt
streamlit==1.31.0
pandas==2.1.4
gspread==5.12.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
Pillow==10.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
reportlab==4.0.7
PyPDF2==3.0.1
